separate tiny generate test from generate test
